Dispatch Statistics visitor to command counter and command length. Delete unnecessary tests.
