+++ src/test/java/edu/kis/powp/jobs2d/TestJobs2dApp.java
@@ -51,9 +51,6 @@ public class TestJobs2dApp {
 	private static void setupCommandTests(Application application) {
 		application.addTest("Load secret command", new SelectLoadSecretCommandOptionListener());
 
-		application.addTest("Draw triangle and count operations", new CommandCounterVisitorTest_drawTriangle());
-		application.addTest("Draw lock and count operations", new ICompoundCommandVisitorTest_drawLock());
-
 		application.addTest("Mouse figure", new SelectMouseFigureOptionListener(application.getFreePanel(), DriverFeature.getDriverManager()));
     
 		application.addTest("Load Macro",new SelectLoadMacroDriverListener());
